Editorial dhanges for PSD04 (#282)
* Editorial dhanges for PSD04

No significant changes other than switching priority of LDP vs OSLC paging. The rest are editorial changes required for publishing the first oslc-op Core PSD.

* Converted additional componts to relative URLs

And fixed shortName to be the same as the file name in order for ReSpec snapshot to have the correct file name (for relative links).

* Set status to WD in order to merge changes to master before PGB vote

After the PG vote, another PR can be used to switch the status to PSD and that merged result will be the tagged release for PSD04.