Add a tool to print a set of shapes
To run this, invoke:

java -cp <sc-classpath> net.open_services.scheck.util.PrintShapes <file
paths, glob patterns, or URIs> ...

and direct the output to a file.

The output is nearly correct Turtle, and may be compared with similar
runs of the tool on earlier sources to verify that you've changed only
what you intended to change regardless of file organization, shape and
property ordering, etc.