Improvements to ShapeChecker:
Improvements to aid with merging multiple shape and/or vocabulary files,
and checking that the results are consistent with the originals -
including the provision of two utility programs to print shapes and
vocabularies in a somewhat canonical manner.

Also allows the vocabulary annotation to declare additional properties
that may be used on terms in a vocabulary - this allows ShapeChecker to
check its own vocabulary without warnings about the severity properties.

Allow multiple -D (or --debug) flags to increase the debug level.

Improvements to character escaping and unescaping.