Sc add shape printer (#358)
* Add a tool to print a set of shapes

To run this, invoke:

java -cp <sc-classpath> net.open_services.scheck.util.PrintShapes <file
paths, glob patterns, or URIs> ...

and direct the output to a file.

The output is nearly correct Turtle, and may be compared with similar
runs of the tool on earlier sources to verify that you've changed only
what you intended to change regardless of file organization, shape and
property ordering, etc.

* Update to run with Java 8 and 14, using Mustache

Needed to update ShapeChecker and hence scapt to run under recent
versions of Java. Due to class library conflicts, new compiler
restrictions, etc, this meant replacing use of Apache Velocity with the
simple Mustache as the template engine for scapt.

* Multiple changes for ShapeChecker:

Add PrintVocabs command
Use updated scapt
Update Gradle to 6.5 to work in Eclipse 2020-06 and Java 14

* Update SpotBugs exclude file for latest scapt and SpotBugs

* Improvements to ShapeChecker tooling

Fix warnings from various linters
Improve shape printer, with support for allowed values
Ensure output from scapt is sorted for consistency across runs

* Add a files to test ShapeChecker PrintShapes tool

* Make PrintShapes produce valid Turtle output

* Improvements to the scapt annotation processor

Allow for additional statements to be added to the ontology
Better handling of characters that need quoting (or not)
Fix some SpotBugs warnings

* Remove obsolete velocity.properties

* Improvements to ShapeChecker:

Improvements to aid with merging multiple shape and/or vocabulary files,
and checking that the results are consistent with the originals -
including the provision of two utility programs to print shapes and
vocabularies in a somewhat canonical manner.

Also allows the vocabulary annotation to declare additional properties
that may be used on terms in a vocabulary - this allows ShapeChecker to
check its own vocabulary without warnings about the severity properties.

Allow multiple -D (or --debug) flags to increase the debug level.

Improvements to character escaping and unescaping.

* Add test shape for canonical shape printer

* Update Gradle build for generated source dir

* Found the right syntax for Gradle to include more resources

This was WAY harder than it should have been. Documentation on Gradle is
seriously imprecise, and the whole Gradle/Kotlin thing really does not
help. This is all over-engineering at its worst!

* Fix out of date path in check-cm.sh, and fail if SC cannot parse files

* Change Cirle-CI scripts to check all domain shapes and vocabs

i.e., don't stop on error

* Run all shape & vocab tests, even if earlier ones failed

Previous attempt using condition: always() in YAML did not work.

* Fix spacing in config.yml

* Fix path error in check-am.sh

* Minor fixes for Checkstyle and similar tools

Co-authored-by: Andrew Berezovskyi <andriib@kth.se>