TRS v3.0 PS01 Editorial Changes (#555)
* These address publishing issues for PSD01

I couldn't find any other issues. The guidance.html document should be moved to the oscl-spec/notes folder when it is completed.  We can address that in PS01.

* More publishing updates

* Ran prettier

* Editorial changes for PS01

Removed the shapes from the TRS spec, and changed unused ChangeEventShape to CreationEventShape and added ModificationEventShape which were missing in the shapes document.

* Added the EBNF for rdfPatch

* Fixed resource shape anchor ids for proper linking

* Changed conformance label prefix to TRS

* Added previous version links

* Back-leveled super-linter to work around failure

* Clarified trs:order semantics in the shapes file